move models/meetup to event namespace
